+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next release
 
 - improvement: add loader for create/delete availability slot
+- improvement: allow admin configure memeber's profile gender/birthday as required
 - Fix a bug: unable to update a space with a deleted machine
 - Fix a bug: unable to get invoice payment details if the account code is same for card/transfer payment method